+++ Libraries/LibTLS/Certificate.cpp
@@ -8,6 +8,7 @@
 #include "Certificate.h"
 #include <AK/Debug.h>
 #include <AK/IPv4Address.h>
+#include <LibCore/DateTime.h>
 #include <LibCrypto/ASN1/ASN1.h>
 #include <LibCrypto/ASN1/DER.h>
 #include <LibCrypto/ASN1/PEM.h>
@@ -906,5 +907,39 @@ ErrorOr<String> RelativeDistinguishedName::to_string() const
 #undef ADD_IF_RECOGNIZED
 
     return cert_name.to_string();
+}
+
+bool Certificate::is_valid() const
+{
+    auto now = UnixDateTime::now();
+
+    if (now < validity.not_before) {
+        dbgln("certificate expired (not yet valid, signed for {})", Core::DateTime::from_timestamp(validity.not_before.seconds_since_epoch()));
+        return false;
+    }
+
+    if (validity.not_after < now) {
+        dbgln("certificate expired (expiry date {})", Core::DateTime::from_timestamp(validity.not_after.seconds_since_epoch()));
+        return false;
+    }
+
+    return true;
+}
+
+// https://www.ietf.org/rfc/rfc5280.html#page-12
+bool Certificate::is_self_signed()
+{
+    if (m_is_self_signed.has_value())
+        return *m_is_self_signed;
+
+    // Self-signed certificates are self-issued certificates where the digital
+    // signature may be verified by the public key bound into the certificate.
+    if (!this->is_self_issued)
+        m_is_self_signed.emplace(false);
+
+    // FIXME: Actually check if we sign ourself
+
+    m_is_self_signed.emplace(true);
+    return *m_is_self_signed;
 }
 }
